dust fit/plots file changes
